+++ zmq_device.c
@@ -62,7 +62,7 @@ static bool s_invoke_device_cb(php_zmq_device_cb_t *cb, uint64_t current_ts)
 		}
 	}
 	if (!Z_ISUNDEF(fc_retval)) {
-		retval = zval_is_true(&fc_retval);
+		retval = zend_is_true(&fc_retval);
 	}
 	zval_ptr_dtor(&params[0]);
 	zval_ptr_dtor(&fc_retval);